Create errors helper
