+++ test/helpers/application_helper_test.rb
@@ -16,16 +16,4 @@ class ApplicationHelperTest < ActiveSupport::TestCase
     assert_match(/\d+\/\d+\/\d+/, time_stamp_for(past))
     assert_match(/\d+\/\d+\/\d+/, time_stamp_for(future))
   end
-
-  test "error_type_from_status returns correct error type" do
-    assert_equal :forbidden, error_type_from_status(403)
-    assert_equal :not_found, error_type_from_status(404)
-    assert_equal :unprocessable, error_type_from_status(422)
-    assert_equal :internal_server, error_type_from_status(500)
-  end
-
-  test "error_type_from_status returns nil for unknown status" do
-    assert_nil error_type_from_status(418) # Teapot status code
-    assert_nil error_type_from_status(999) # Arbitrary unknown status
-  end
 end
